Small change to project Servlets
The class HelloWorldServlet was thread safe but we made it unsafe by
adding the attribute visits. If there are no class variables declared
then the class is thread safe as there is no data to be shared between
threads.
The field visits is a list that adds a new date in the class method and
we have surrounded it with a synchronized block to keep it thread safe